Code Cleanup (#358)

+++ src/minecraft/universalelectricity/components/client/gui/GuiBatteryBox.java
@@ -63,7 +63,7 @@ public class GuiBatteryBox extends GuiContainer
 		// Background energy bar
 		this.drawTexturedModalRect(containerWidth, containerHeight, 0, 0, this.xSize, this.ySize);
 		// Foreground energy bar
-		int scale = (int) (((double) this.tileEntity.getJoules() / this.tileEntity.getMaxJoules()) * 72);
+		int scale = (int) ((this.tileEntity.getJoules() / this.tileEntity.getMaxJoules()) * 72);
 		this.drawTexturedModalRect(containerWidth + 87, containerHeight + 52, 176, 0, scale, 20);
 	}
 }
