+++ src/minecraft/universalelectricity/components/common/block/BlockBasicMachine.java
@@ -68,31 +68,31 @@ public class BlockBasicMachine extends BlockAdvanced
 			if (tileEntity.generateWatts > 0)
 			{
 				int metadata = par1World.getBlockMetadata(x, y, z);
-				float var7 = (float) x + 0.5F;
-				float var8 = (float) y + 0.0F + par5Random.nextFloat() * 6.0F / 16.0F;
-				float var9 = (float) z + 0.5F;
+				float var7 = x + 0.5F;
+				float var8 = y + 0.0F + par5Random.nextFloat() * 6.0F / 16.0F;
+				float var9 = z + 0.5F;
 				float var10 = 0.52F;
 				float var11 = par5Random.nextFloat() * 0.6F - 0.3F;
 
 				if (metadata == 3)
 				{
-					par1World.spawnParticle("smoke", (double) (var7 - var10), (double) var8, (double) (var9 + var11), 0.0D, 0.0D, 0.0D);
-					par1World.spawnParticle("flame", (double) (var7 - var10), (double) var8, (double) (var9 + var11), 0.0D, 0.0D, 0.0D);
+					par1World.spawnParticle("smoke", var7 - var10, var8, var9 + var11, 0.0D, 0.0D, 0.0D);
+					par1World.spawnParticle("flame", var7 - var10, var8, var9 + var11, 0.0D, 0.0D, 0.0D);
 				}
 				else if (metadata == 2)
 				{
-					par1World.spawnParticle("smoke", (double) (var7 + var10), (double) var8, (double) (var9 + var11), 0.0D, 0.0D, 0.0D);
-					par1World.spawnParticle("flame", (double) (var7 + var10), (double) var8, (double) (var9 + var11), 0.0D, 0.0D, 0.0D);
+					par1World.spawnParticle("smoke", var7 + var10, var8, var9 + var11, 0.0D, 0.0D, 0.0D);
+					par1World.spawnParticle("flame", var7 + var10, var8, var9 + var11, 0.0D, 0.0D, 0.0D);
 				}
 				else if (metadata == 1)
 				{
-					par1World.spawnParticle("smoke", (double) (var7 + var11), (double) var8, (double) (var9 - var10), 0.0D, 0.0D, 0.0D);
-					par1World.spawnParticle("flame", (double) (var7 + var11), (double) var8, (double) (var9 - var10), 0.0D, 0.0D, 0.0D);
+					par1World.spawnParticle("smoke", var7 + var11, var8, var9 - var10, 0.0D, 0.0D, 0.0D);
+					par1World.spawnParticle("flame", var7 + var11, var8, var9 - var10, 0.0D, 0.0D, 0.0D);
 				}
 				else if (metadata == 0)
 				{
-					par1World.spawnParticle("smoke", (double) (var7 + var11), (double) var8, (double) (var9 + var10), 0.0D, 0.0D, 0.0D);
-					par1World.spawnParticle("flame", (double) (var7 + var11), (double) var8, (double) (var9 + var10), 0.0D, 0.0D, 0.0D);
+					par1World.spawnParticle("smoke", var7 + var11, var8, var9 + var10, 0.0D, 0.0D, 0.0D);
+					par1World.spawnParticle("flame", var7 + var11, var8, var9 + var10, 0.0D, 0.0D, 0.0D);
 				}
 			}
 		}
